Refactor invoice handling and payment processing; enhance HR dashboard with aggregated statistics; improve sales stock management with pagination and status updates
- Updated invoice schema to handle cheque numbers as strings.
- Modified SalesService to back-fill invoice numbers and handle cheque payments more robustly.
- Added methods for branch access validation and invoice total recomputation.
- Enhanced HR dashboard to fetch and display aggregated statistics from the server.
- Simplified API calls in HR dashboard by consolidating data fetching.
- Implemented paginated sales stock API with summary statistics.
- Updated sales stock dashboard to reflect new paginated data structure and improved filtering.
- Refined stock status handling and added new status options for better clarity.

+++ backend/app/modules/purchasing/schemas.py
@@ -29,7 +29,7 @@ class SupplierBase(BaseModel):
     home_contact_number: Optional[str] = None
     mobile_contact_number: str
     credit_days: int
-    max_credit_limit: int
+    max_credit_limit: Decimal
     active: bool = True
     country_id: Optional[int] = None
 
@@ -59,15 +59,15 @@ class SupplierUpdate(BaseModel):
     home_contact_number: Optional[str] = None
     mobile_contact_number: Optional[str] = None
     credit_days: Optional[int] = None
-    max_credit_limit: Optional[int] = None
+    max_credit_limit: Optional[Decimal] = None
     active: Optional[bool] = None
     country_id: Optional[int] = None
 
 class Supplier(SupplierBase, TijaeroBaseSchema):
     id: int
     date_joined: datetime
-    left_credit_amount: Optional[int] = None
-    initial_credit_amount: Optional[int] = None
+    left_credit_amount: Optional[Decimal] = None
+    initial_credit_amount: Optional[Decimal] = None
 
 class PurchasingOrderItemBase(BaseModel):
     product_id: int
